+++ packages/runtime-js/src/builtins/index.ts
@@ -2,6 +2,7 @@ import { registerBuiltin } from '@rifty/io';
 import assertModule, { strict as assertStrict } from './assert.ts';
 import bufferModule, { Buffer } from './buffer.ts';
 import childProcessModule from './child_process.ts';
+import consoleModule from './console.ts';
 import cryptoModule from './crypto.ts';
 import diagnosticsChannelModule from './diagnostics_channel.ts';
 /**
@@ -87,6 +88,7 @@ registerBuiltin('child_process', () => childProcessModule);
 registerBuiltin('worker_threads', () => workerThreadsModule);
 registerBuiltin('os', () => osModule);
 registerBuiltin('crypto', () => cryptoModule);
+registerBuiltin('console', () => consoleModule);
 registerBuiltin('diagnostics_channel', () => diagnosticsChannelModule);
 registerBuiltin('module', () => moduleModule);
 registerBuiltin('perf_hooks', () => perfHooksModule);